LayoutTests/imported/w3c/web-platform-tests/css/css-transforms/animation/composited-transform-expected.txt


FAIL An additive transform animation on-top of a replace transform animation should composite correctly assert_equals: expected "matrix(1, 0, 0, 1, 20, 80)" but got "matrix(1, 0, 0, 1, 0, 80)"
